requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=0.12.0=pypi_0
animation=0.0.7=pypi_0
arviz=0.11.2=pyhd8ed1ab_1
autopep8=1.5.6=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2020.12.5=h5b45459_0
cachetools=4.2.2=pypi_0
certifi=2020.12.5=py39hcbf5309_1
cffi=1.14.5=py39hcd4344a_0
cftime=1.2.1=py39h8512cfa_1
chardet=4.0.0=py39haa95532_1003
cloudpickle=1.6.0=py_0
cmdstanpy=0.9.5=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
convertdate=2.3.2=pypi_0
cryptography=3.4.7=py39h71e12ea_0
cudatoolkit=11.1.1=heb2d755_7
curl=7.76.1=h789b8ee_2
cycler=0.10.0=py39haa95532_0
cython=0.29.23=py39h415ef7b_0
decorator=4.4.2=pyhd3eb1b0_0
ephem=3.7.7.1=pypi_0
fbprophet=0.7.1=pypi_0
ffmpeg=4.3.1=ha925a31_0
freetype=2.10.4=hd328e21_0
future=0.18.2=py39hcbf5309_3
google-auth=1.30.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
grpcio=1.37.1=pypi_0
gym=0.18.0=py39h832f523_0
hdf4=4.2.13=h712560f_2
hdf5=1.10.6=nompi_h5268f04_1114
hijri-converter=2.1.1=pypi_0
holidays=0.11.1=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
intel-openmp=2020.2=254
ipykernel=5.5.0=pypi_0
ipython=7.21.0=py39hd4e2768_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.2=py39haa95532_1
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=hb83a4c4_2
jupyter_client=6.1.7=py_0
jupyter_core=4.7.1=py39haa95532_0
kaggle=1.5.12=pypi_0
kiwisolver=1.3.1=py39hd77b12b_0
korean-lunar-calendar=0.2.1=pypi_0
krb5=1.19.1=hbae68bd_0
libcurl=7.76.1=h789b8ee_2
libnetcdf=4.6.1=hf59b723_3
libpng=1.6.37=h2a8f88b_0
libpython=2.0=py39hcbf5309_1
libsodium=1.0.18=h62dcd97_0
libssh2=1.9.0=h680486a_6
libtiff=4.1.0=h56a325e_1
libuv=1.40.0=he774522_0
lunarcalendar=0.0.9=pypi_0
lz4-c=1.9.3=h2bbff1b_0
m2w64-binutils=2.25.1=5
m2w64-bzip2=1.0.6=6
m2w64-crt-git=5.0.0.4636.2595836=2
m2w64-gcc=5.3.0=6
m2w64-gcc-ada=5.3.0=6
m2w64-gcc-fortran=5.3.0=6
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gcc-objc=5.3.0=6
m2w64-gmp=6.1.0=2
m2w64-headers-git=5.0.0.4636.c0ad18a=2
m2w64-isl=0.16.1=2
m2w64-libiconv=1.14=6
m2w64-libmangle-git=5.0.0.4509.2e5a9a2=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
m2w64-make=4.1.2351.a80a8b8=2
m2w64-mpc=1.0.3=3
m2w64-mpfr=3.1.4=4
m2w64-pkg-config=0.29.1=2
m2w64-toolchain=5.3.0=7
m2w64-tools-git=5.0.0.4592.90b8472=2
m2w64-windows-default-manifest=6.4=3
m2w64-winpthreads-git=5.0.0.4634.697f757=2
m2w64-zlib=1.2.8=10
markdown=3.3.4=pypi_0
matplotlib=3.3.4=py39haa95532_0
matplotlib-base=3.3.4=py39h49ac443_0
mkl=2020.2=256
mkl-service=2.3.0=py39h196d8e1_0
mkl_fft=1.3.0=py39h46781fe_0
mkl_random=1.0.2=py39h848d8c7_0
msys2-conda-epoch=20160418=1
netcdf4=1.5.6=py39hb33177a_0
ninja=1.10.2=py39h6d14046_0
numpy=1.19.2=py39h729668d_0
numpy-base=1.19.2=py39hbd0edd7_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py_0
openssl=1.1.1k=h8ffe710_0
packaging=20.9=pyh44b312d_0
pandas=1.2.4=py39hd77b12b_0
parso=0.7.0=py_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py39h4fa10fc_0
pip=21.0.1=py39haa95532_0
prompt-toolkit=3.0.8=py_0
protobuf=3.15.8=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pychalk=2.0.1=pypi_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.20=py_2
pyglet=1.5.0=pypi_0
pygments=2.8.0=pyhd3eb1b0_0
pymeeus=0.5.11=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39hd77b12b_6
pysocks=1.7.1=py39haa95532_0
pystan=2.19.1.1=py39hac22706_0
python=3.9.1=h6244533_2
python-dateutil=2.8.1=pyhd3eb1b0_0
python-slugify=5.0.0=pypi_0
python_abi=3.9=1_cp39
pytorch=1.8.0=py3.9_cuda11.1_cudnn8_0
pytz=2021.1=pyhd3eb1b0_0
pywin32=228=py39he774522_0
pyzmq=20.0.0=py39hd77b12b_1
qt=5.9.7=vc14h73c81de_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.1=py39hf11a4ad_0
scipy=1.6.1=py39h14eb087_0
setuptools=52.0.0=py39haa95532_0
setuptools-git=1.2=pypi_0
sip=4.19.13=py39hd77b12b_0
six=1.15.0=py39haa95532_0
sqlite=3.33.0=h2a8f88b_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
text-unidecode=1.3=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
toml=0.10.2=pyhd3eb1b0_0
torch-tb-profiler=0.1.0=pypi_0
torchaudio=0.8.0=py39
torchvision=0.9.0=py39_cu111
tornado=6.1=py39h2bbff1b_0
tqdm=4.60.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.4=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39h2bbff1b_0
xarray=0.18.0=pyhd8ed1ab_0
xz=5.2.5=h62dcd97_0
zeromq=4.3.3=ha925a31_3
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
